move data fetching to separate file
